Create DI extension, compiler pass and controller
The compiler pass is responsible for extracting parameters from asset package services. Due to a limited public API, we cannot extract parameters after service construction without resorting to reflection.

The controller is an extremely lightweight class, which will simply provide extracted parameters to a frontend JavaScript library, as is done in FOSJsRoutingBundle.